Make parameters comply with fcitx5's gclient lib

+++ src/fcitx-fbterm.cpp
@@ -255,7 +255,8 @@ static void process_raw_key(char *buf, unsigned int len)
 
             fcitx_g_client_focus_in(client);
 
-            if (keysym == FcitxKey_None || fcitx_g_client_process_key_sync(client, keysym, code, state, down ? FALSE : TRUE, 0) <= 0) {
+            if (keysym == FcitxKey_None || fcitx_g_client_process_key_sync(client, keysym, code,
+                                                                           static_cast<guint32>(state), down ? FALSE : TRUE, 0) <= 0) {
                 char *str = keysym_to_term_string(linux_keysym, down);
                 if (str)
                     put_im_text(str, strlen(str));
